requirements.txt
blinker==1.9.0
brotli==1.2.0
cachelib==0.13.0
click==8.3.1
colorama==0.4.6
flask==3.1.2
flask-caching==2.3.1
flask-squeeze==3.5.0
flask-wtf==1.2.2
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.3
python-dotenv==1.2.1
rcssmin==1.2.2
rjsmin==1.2.5
werkzeug==3.1.3
wtforms==3.2.1
